BTCD Block Listing implementation
Summary:
Replace listing by timestamps to filter by heights.
Current code uses getblockhashes a bitpay/bitcoind custom method which filters by timestamps.
This method is used on the explorer to filter the blocks by dates, but we don't need and and we also are not  using bitcoind.
Since btcd doesn't have that method, we require our way to get the latest blocks, so we can easily do that by getting the lastest block
using the getbestblockhash rpc method, then we just need to get the X(limit) number of blocks above that one. We are using simple
math by substracting 1 to the block height. Using previous hash can be done, but implementation requires much more effort.
In this way we reuse tone of code, adding just a small piece.

Apply this patch on the insight-api

Reviewers: barath

Reviewed By: barath

Differential Revision: https://phabricator.bitgo.com/D3990